[REFACTOR] response 폴더명 dto로 변경
