+++ src/main/java/gongalgongal/gongalgongal_spring/dto/UserSignupRequest.java
@@ -1,4 +1,4 @@
-package gongalgongal.gongalgongal_spring.response;
+package gongalgongal.gongalgongal_spring.dto;
 
 import lombok.Data;
 import lombok.NoArgsConstructor;
