+++ src/main/java/gongalgongal/gongalgongal_spring/service/AuthService.java
@@ -2,9 +2,9 @@ package gongalgongal.gongalgongal_spring.service;
 
 import gongalgongal.gongalgongal_spring.model.User;
 import gongalgongal.gongalgongal_spring.repository.UserRepository;
-import gongalgongal.gongalgongal_spring.response.AuthResponse;
-import gongalgongal.gongalgongal_spring.response.UserLoginRequest;
-import gongalgongal.gongalgongal_spring.response.UserSignupRequest;
+import gongalgongal.gongalgongal_spring.dto.AuthResponse;
+import gongalgongal.gongalgongal_spring.dto.UserLoginRequest;
+import gongalgongal.gongalgongal_spring.dto.UserSignupRequest;
 import gongalgongal.gongalgongal_spring.util.JwtUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
